+++ src/Pumukit/SchemaBundle/Security/Permission.php
@@ -41,6 +41,9 @@ class Permission
     public const DISABLED_TRACK_PROFILES = 'ROLE_DISABLED_WIZARD_TRACK_PROFILES';
     public const DISABLED_TRACK_PRIORITY = 'ROLE_DISABLED_WIZARD_TRACK_PRIORITY';
     public const ADD_EXTERNAL_PLAYER = 'ROLE_ADD_EXTERNAL_PLAYER';
+    public const AUTO_CREATE_PERSONAL_SERIES = 'ROLE_AUTO_CREATE_PERSONAL_SERIES';
+    public const ACCESS_HEAD_AND_TAIL_MANAGER = 'ROLE_ACCESS_HEAD_AND_TAIL_MANAGER';
+    public const ADD_HEAD_AND_TAIL = 'ROLE_ADD_HEAD_AND_TAIL';
 
     // Permissions for series metadata inputs
     public const ACCESS_SERIES_META_LAST_ANNOUNCES = 'ROLE_ACCESS_SERIES_META_LAST_ANNOUNCES';
@@ -300,6 +303,27 @@ class Permission
         ],
         self::ADD_EXTERNAL_PLAYER => [
             'description' => 'Add an external player (iframe) into a multimedia object',
+            'dependencies' => [
+                PermissionProfile::SCOPE_GLOBAL => [],
+                PermissionProfile::SCOPE_PERSONAL => [],
+            ],
+        ],
+        self::AUTO_CREATE_PERSONAL_SERIES => [
+            'description' => 'Auto create personal series for the user',
+            'dependencies' => [
+                PermissionProfile::SCOPE_GLOBAL => [],
+                PermissionProfile::SCOPE_PERSONAL => [],
+            ],
+        ],
+        self::ACCESS_HEAD_AND_TAIL_MANAGER => [
+            'description' => 'Allow access to head and tail video manager',
+            'dependencies' => [
+                PermissionProfile::SCOPE_GLOBAL => [],
+                PermissionProfile::SCOPE_PERSONAL => [],
+            ],
+        ],
+        self::ADD_HEAD_AND_TAIL => [
+            'description' => 'Allow add head and tail video on multimedia objects and series',
             'dependencies' => [
                 PermissionProfile::SCOPE_GLOBAL => [],
                 PermissionProfile::SCOPE_PERSONAL => [],