+++ src/Pumukit/StatsBundle/Services/StatsService.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 namespace Pumukit\StatsBundle\Services;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
@@ -284,6 +282,12 @@ class StatsService
             ],
             ['$out' => 'ViewsAggregation'],
         ];
+
+//        $result = $viewsLogColl->aggregate($pipeline, ['cursor' => [], 'allowDiskUse' => true]);
+
+        // >>> TTK-25016 do not remove the next line.
+//        $workAround = \count($result);
+        // <<< TTK-25016 do not remove the above line.
 
         return $viewsLogColl->aggregate($pipeline, ['cursor' => [], 'allowDiskUse' => true]);
     }